fix response delete

+++ lib/tasks/destroy_by_brand.rake
@@ -15,7 +15,7 @@ namespace :destroy_by_brand do
       insales_product_id = get_id_var(fid)
 
       response = Services::DeleteProductInsales.new(insales_product_id).call
-      if response["status"] == 'ok'
+      if response.present? && response["status"] == 'ok'
         product_app.destroy
       end
     end
